Revert "Finish modification"
This reverts commit 6281abf001d8b7425b9c046d3ccda197e69766b9.

+++ bitArray.h
@@ -20,7 +20,7 @@ public:
     bool set(int index);
     void reset(int start,int end);
     void print();
-    void randomMask(float  r);
+    void randomMask(double r);
     void randomMask(int r);
     virtual ~bitArray();
 private:
